template_python_test_repo/pyproject.toml
[tool.poetry]

name = "template_python_test_repo"
version = "0.0"
description = "Sample project generated from Jace's Python Template."

packages = [{ include = "template_python_test_repo" }]

license = "MIT"
authors = ["Aleksei Grachev <aljeshishe@gmail.com>"]

readme = "README.md"
homepage = "https://pypi.org/project/template_python_test_repo"
documentation = "https://template_python_test_repo.readthedocs.io"
repository = "https://github.com/aljeshishe/template_python_test_repo"

keywords = [
]
classifiers = [
    # TODO: update this list to match your application: https://pypi.org/pypi?%3Aaction=list_classifiers
    "Development Status :: 1 - Planning",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]

[tool.poetry.dependencies]

python = "^3.11"

# TODO: Remove these and add your library's requirements
click = "*"
minilog = "*"
requests = "^2.28.2"
scrapy = "^2.7.1"
jsonpath-ng = "^1.5.3"
pandas = "^1.5.3"
jupyter = "^1.0.0"
rootpath = "^0.1.1"
kaleido = "0.2.1"
dtale = "^2.11.0"
pandas-utils = {git = "https://github.com/aljeshishe/pandas_utils"}

[tool.poetry.dev-dependencies]

# Formatters
black = "^22.1"
tomli = "*" # missing 'black' dependency
isort = "^5.10"

# Linters
mypy = "^1.0"
pydocstyle = "^6.1"
pylint = "~2.15"
wrapt = "*" # missing 'pylint' dependency

# Testing
pytest = "^7.1"
pytest-describe = "^2.0"
pytest-expecter = "^3.0"
pytest-random = "*"
pytest-cov = "^3.0"
freezegun = "*"

# Reports
coveragespace = "^6.0"

# Documentation
mkdocs = "~1.3"
pygments = "^2.11.1"

# Tooling
pyinstaller = "*"
sniffer = "*"
MacFSEvents = { version = "*", platform = "darwin" }
pync = { version = "*", platform = "darwin" }
ipython = "^7.12.0"

[tool.poetry.scripts]

template_python_test_repo = "template_python_test_repo.cli:main"

[tool.black]

quiet = true

[tool.isort]

profile = "black"

[tool.mypy]

ignore_missing_imports = true
no_implicit_optional = true
check_untyped_defs = true

cache_dir = ".cache/mypy/"

[tool.pytest.ini_options]

addopts = """
--strict-markers

-r sxX
--show-capture=log

--cov-report=html
--cov-report=term-missing:skip-covered
--no-cov-on-fail
"""

cache_dir = ".cache/pytest/"

markers = []

[build-system]

requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
